move plugin manager to plugin folder
