+++ src/app/plugin/pluginAPI.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = {
+
+}
