+++ src/app/plugin/pluginManager.js
